Refactoring SigV4 signer to new class + tests
