+++ mr/src/main/java/org/opensearch/hadoop/rest/commonshttp/CommonsHttpTransport.java
@@ -45,6 +45,7 @@ import org.opensearch.hadoop.rest.ReusableInputStream;
 import org.opensearch.hadoop.rest.SimpleResponse;
 import org.opensearch.hadoop.rest.Transport;
 import org.opensearch.hadoop.rest.commonshttp.auth.OpenSearchHadoopAuthPolicies;
+import org.opensearch.hadoop.rest.commonshttp.auth.aws.AwsV4Signer;
 import org.opensearch.hadoop.rest.commonshttp.auth.bearer.OpenSearchApiKeyAuthScheme;
 import org.opensearch.hadoop.rest.commonshttp.auth.bearer.OpenSearchApiKeyCredentials;
 import org.opensearch.hadoop.rest.commonshttp.auth.spnego.SpnegoAuthScheme;
@@ -89,56 +90,21 @@ import org.opensearch.hadoop.util.ReflectionUtils;
 import org.opensearch.hadoop.util.StringUtils;
 import org.opensearch.hadoop.util.encoding.HttpEncodingTools;
 
-import org.opensearch.hadoop.thirdparty.amazon.awssdk.auth.credentials.AwsCredentialsProvider;
-import org.opensearch.hadoop.thirdparty.amazon.awssdk.auth.credentials.DefaultCredentialsProvider;
-import org.opensearch.hadoop.thirdparty.amazon.awssdk.auth.signer.Aws4Signer;
-import org.opensearch.hadoop.thirdparty.amazon.awssdk.auth.signer.params.Aws4SignerParams;
-import org.opensearch.hadoop.thirdparty.amazon.awssdk.regions.Region;
-import org.opensearch.hadoop.thirdparty.amazon.awssdk.http.SdkHttpFullRequest;
-import org.opensearch.hadoop.thirdparty.amazon.awssdk.http.SdkHttpClient;
-import org.opensearch.hadoop.thirdparty.amazon.awssdk.http.SdkHttpMethod;
-import org.opensearch.hadoop.thirdparty.amazon.awssdk.http.SdkHttpResponse;
-import org.opensearch.hadoop.thirdparty.amazon.awssdk.utils.http.SdkHttpUtils;
 
 import javax.security.auth.kerberos.KerberosPrincipal;
 import java.io.ByteArrayInputStream;
 import java.io.Closeable;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.StringBufferInputStream;
-import java.io.UnsupportedEncodingException;
 import java.lang.reflect.Method;
 import java.net.Socket;
-import java.net.URLDecoder;
-import java.nio.charset.StandardCharsets;
 import java.security.PrivilegedExceptionAction;
-import java.time.LocalDateTime;
-import java.time.ZoneOffset;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
-import org.opensearch.hadoop.thirdparty.google.common.base.Optional;
-import org.opensearch.hadoop.thirdparty.google.common.base.Splitter;
-import org.opensearch.hadoop.thirdparty.google.common.base.Strings;
-import org.opensearch.hadoop.thirdparty.google.common.base.Supplier;
-import org.opensearch.hadoop.thirdparty.google.common.collect.ImmutableListMultimap;
-import org.opensearch.hadoop.thirdparty.google.common.collect.Multimap;
-import org.opensearch.hadoop.thirdparty.google.common.collect.ImmutableMap;
-import org.opensearch.hadoop.thirdparty.google.common.collect.ImmutableList;
-import org.opensearch.hadoop.thirdparty.google.common.base.Joiner;
 import java.io.ByteArrayOutputStream;
-import java.net.URLDecoder;
-import java.nio.charset.StandardCharsets;
-import java.time.LocalDateTime;
-import java.time.ZoneOffset;
-import java.util.Collections;
-import java.net.URISyntaxException;
-import java.net.URLEncoder;
-import java.util.TreeMap;
-import java.util.Collection;
 
 /**
  * Transport implemented on top of Commons Http. Provides transport retries.
@@ -170,7 +136,6 @@ public class CommonsHttpTransport implements Transport, StatsAware {
     private UserProvider proxyUserProvider = null;
     private String runAsUser = null;
 
-    private static final Joiner AMPERSAND_JOINER = Joiner.on('&');
 
     /** If the HTTP Connection is made through a proxy */
     private boolean isProxied = false;
@@ -763,7 +728,8 @@ public class CommonsHttpTransport implements Transport, StatsAware {
         }
 
         if (settings.getAwsSigV4Enabled()) {
-            awsSigV4SignRequest(request, http, bodyBytes);
+            AwsV4Signer awsV4Signer = new AwsV4Signer(settings, httpInfo);
+            awsV4Signer.sign(request, http, bodyBytes);
         }
 
         if (executingProvider != null) {
@@ -798,86 +764,7 @@ public class CommonsHttpTransport implements Transport, StatsAware {
         return new SimpleResponse(http.getStatusCode(), new ResponseInputStream(http), httpInfo, headers);
     }
 
-    private String queryParamsString(Multimap<String, String> queryParams) {
-        final ImmutableList.Builder<String> result = ImmutableList.builder();
-        for (Map.Entry<String, Collection<String>> param : new TreeMap<>(queryParams.asMap()).entrySet()) {
-            for (String value : param.getValue()) {
-                result.add(SdkHttpUtils.urlEncode(param.getKey()) + '=' + SdkHttpUtils.urlEncode(value));
-            }
-        }
 
-        return '?' + AMPERSAND_JOINER.join(result.build());
-    }
-
-    private void awsSigV4SignRequest(Request request, HttpMethod http, byte[] bodyBytes)
-            throws UnsupportedEncodingException {
-        String awsRegion = settings.getAwsSigV4Region();
-        String awsServiceName = settings.getAwsSigV4ServiceName();
-
-        Region signingRegion = Region.of(awsRegion);
-
-        final AwsCredentialsProvider credentials = DefaultCredentialsProvider.create();
-
-        Aws4SignerParams signerParams = Aws4SignerParams.builder()
-                .awsCredentials(credentials.resolveCredentials())
-                .signingName(awsServiceName)
-                .signingRegion(signingRegion)
-                .build();
-
-        SdkHttpFullRequest.Builder req = SdkHttpFullRequest.builder()
-                .method(SdkHttpMethod.fromValue(request.method().name()));
-
-        StringBuilder url = new StringBuilder();
-        url.append(httpInfo);
-        String path = request.path().toString();
-        if (!path.startsWith("/")) {
-            url.append('/');
-        }
-        url.append(path);
-
-        Splitter queryStringSplitter = Splitter.on('&').trimResults().omitEmptyStrings();
-        final Iterable<String> rawParams = request.params() != null ? queryStringSplitter.split(request.params())
-                : Collections.emptyList();
-
-        final ImmutableListMultimap.Builder<String, String> queryParams = ImmutableListMultimap.builder();
-
-        for (String rawParam : rawParams) {
-            if (!Strings.isNullOrEmpty(rawParam)) {
-                final String pair = URLDecoder.decode(rawParam, StandardCharsets.UTF_8.name());
-                final int index = pair.indexOf('=');
-                if (index > 0) {
-                    final String key = pair.substring(0, index);
-                    final String value = pair.substring(index + 1);
-                    queryParams.put(key, value);
-                } else {
-                    queryParams.put(pair, "");
-                }
-            }
-        }
-
-        String params = queryParamsString(queryParams.build());
-
-        url.append(params);
-
-        try {
-            req.uri(new java.net.URI(url.toString()));
-        } catch (URISyntaxException e) {
-            throw new IllegalArgumentException("Invalid request URI: " + request.uri().toString());
-        }
-
-        if (request.body() != null) {
-            req.contentStreamProvider(() -> new ByteArrayInputStream(bodyBytes));
-        }
-        req.putHeader("x-amz-content-sha256", "required");
-
-        SdkHttpFullRequest signedRequest = Aws4Signer.create().sign(req.build(), signerParams);
-
-        final ImmutableMap.Builder<String, String> signerHeaders = ImmutableMap.builder();
-
-        for (Map.Entry<String, List<String>> entry : signedRequest.headers().entrySet()) {
-            http.setRequestHeader(entry.getKey(), entry.getValue().get(0));
-        }
-    }
 
     /**
      * Actually perform the request
